+++ test/irb/test_context.rb
@@ -42,6 +42,17 @@ module TestIRB
       IRB.conf[:VERBOSE] = false
       workspace = IRB::WorkSpace.new(Object.new)
       @context = IRB::Context.new(nil, workspace, TestInputMethod.new)
+
+      @get_screen_size = Reline.method(:get_screen_size)
+      Reline.instance_eval { undef :get_screen_size }
+      def Reline.get_screen_size
+        [36, 80]
+      end
+    end
+
+    def teardown
+      Reline.instance_eval { undef :get_screen_size }
+      Reline.define_singleton_method(:get_screen_size, @get_screen_size)
     end
 
     def test_last_value
@@ -103,6 +114,25 @@ module TestIRB
                            /\s*/], out)
     ensure
       $VERBOSE = verbose
+    end
+
+    def test_eval_input_raise2x
+      skip if RUBY_ENGINE == 'truffleruby'
+      input = TestInputMethod.new([
+        "raise 'Foo'\n",
+        "raise 'Bar'\n",
+        "_\n",
+      ])
+      irb = IRB::Irb.new(IRB::WorkSpace.new(Object.new), input)
+      out, err = capture_output do
+        irb.eval_input
+      end
+      assert_empty err
+      assert_pattern_list([
+          :*, /\(irb\):1:in `<main>': Foo \(RuntimeError\)\n/,
+          :*, /\(irb\):2:in `<main>': Bar \(RuntimeError\)\n/,
+          :*, /#<RuntimeError: Bar>\n/,
+        ], out)
     end
 
     def test_eval_object_without_inspect_method
@@ -428,7 +458,7 @@ module TestIRB
         irb.eval_input
       end
       assert_empty err
-      if '2.5.0' <= RUBY_VERSION && RUBY_VERSION < '3.0.0'
+      if '2.5.0' <= RUBY_VERSION && RUBY_VERSION < '3.0.0' && STDOUT.tty?
         expected = [
           :*, /Traceback \(most recent call last\):\n/,
           :*, /\t 2: from \(irb\):1:in `<main>'\n/,
@@ -458,7 +488,7 @@ module TestIRB
         irb.eval_input
       end
       assert_empty err
-      if '2.5.0' <= RUBY_VERSION && RUBY_VERSION < '3.0.0'
+      if '2.5.0' <= RUBY_VERSION && RUBY_VERSION < '3.0.0' && STDOUT.tty?
         expected = [
           :*, /Traceback \(most recent call last\):\n/,
           :*, /\t 2: from \(irb\):1:in `<main>'\n/,
@@ -494,7 +524,7 @@ module TestIRB
         irb.eval_input
       end
       assert_empty err
-      if '2.5.0' <= RUBY_VERSION && RUBY_VERSION < '3.0.0'
+      if '2.5.0' <= RUBY_VERSION && RUBY_VERSION < '3.0.0' && STDOUT.tty?
         expected = [
           :*, /Traceback \(most recent call last\):\n/,
           :*, /\t... 5 levels...\n/,