+++ test/ruby/test_time_tz.rb
@@ -606,6 +606,8 @@ module TestTimeTZ::WithTZ
     assert_equal(time_class.utc(2018, 9, 1, 12+h, m, 0).to_i, t.to_i)
     assert_equal(6, t.wday)
     assert_equal(244, t.yday)
+    assert_equal(t, time_class.new(2018, 9, 1, 12, in: tzarg))
+    assert_raise(ArgumentError) {time_class.new(2018, 9, 1, 12, 0, 0, tzarg, in: tzarg)}
   end
 
   def subtest_now(time_class, tz, tzarg, tzname, abbr, utc_offset)