+++ vm_method.c
@@ -128,7 +128,7 @@ rb_vm_cc_invalidate(const struct rb_callcache *cc)
 
 
 static inline void
-vm_me_invalidate_cache(rb_callable_method_entry_t *cme)
+vm_cme_invalidate(rb_callable_method_entry_t *cme)
 {
     VM_ASSERT(IMEMO_TYPE_P(cme, imemo_ment));
     VM_ASSERT(callable_method_entry_p(cme));
@@ -143,6 +143,21 @@ rb_clear_constant_cache(void)
     INC_GLOBAL_CONSTANT_STATE();
 }
 
+static void
+invalidate_negative_cache(ID mid, bool invalidate_cme)
+{
+    const rb_callable_method_entry_t *cme;
+    rb_vm_t *vm = GET_VM();
+
+    if (rb_id_table_lookup(vm->negative_cme_table, mid, (VALUE *)&cme)) {
+        rb_id_table_delete(vm->negative_cme_table, mid);
+        if (invalidate_cme) {
+            vm_cme_invalidate((rb_callable_method_entry_t *)cme);
+            RB_DEBUG_COUNTER_INC(cc_invalidate_negative);
+        }
+    }
+}
+
 static rb_method_entry_t *rb_method_entry_alloc(ID called_id, VALUE owner, VALUE defined_class, const rb_method_definition_t *def);
 const rb_method_entry_t * rb_method_entry_clone(const rb_method_entry_t *src_me);
 static const rb_callable_method_entry_t *complemented_callable_method_entry(VALUE klass, ID id);
@@ -151,6 +166,7 @@ static void
 clear_method_cache_by_id_in_class(VALUE klass, ID mid)
 {
     VM_ASSERT(RB_TYPE_P(klass, T_CLASS) || RB_TYPE_P(klass, T_ICLASS));
+    if (rb_objspace_garbage_object_p(klass)) return;
 
     if (LIKELY(RCLASS_EXT(klass)->subclasses == NULL)) {
         // no subclasses
@@ -161,6 +177,7 @@ clear_method_cache_by_id_in_class(VALUE klass, ID mid)
 
         // invalidate CCs
         if (cc_tbl && rb_id_table_lookup(cc_tbl, mid, (VALUE *)&ccs)) {
+            if (NIL_P(ccs->cme->owner)) invalidate_negative_cache(mid, false);
             rb_vm_ccs_free(ccs);
             rb_id_table_delete(cc_tbl, mid);
             RB_DEBUG_COUNTER_INC(cc_invalidate_leaf_ccs);
@@ -192,7 +209,7 @@ clear_method_cache_by_id_in_class(VALUE klass, ID mid)
                     rb_method_table_insert(origin, RCLASS_M_TBL(origin), mid, new_cme);
                 }
 
-                vm_me_invalidate_cache((rb_callable_method_entry_t *)cme);
+                vm_cme_invalidate((rb_callable_method_entry_t *)cme);
                 RB_DEBUG_COUNTER_INC(cc_invalidate_tree_cme);
             }
 
@@ -209,13 +226,7 @@ clear_method_cache_by_id_in_class(VALUE klass, ID mid)
             RB_DEBUG_COUNTER_INC(cc_invalidate_tree);
         }
         else {
-            rb_vm_t *vm = GET_VM();
-            if (rb_id_table_lookup(vm->negative_cme_table, mid, (VALUE *)&cme)) {
-                rb_id_table_delete(vm->negative_cme_table, mid);
-                vm_me_invalidate_cache((rb_callable_method_entry_t *)cme);
-
-                RB_DEBUG_COUNTER_INC(cc_invalidate_negative);
-            }
+            invalidate_negative_cache(mid, true);
         }
     }
 }