+++ NEWS.md
@@ -21,6 +21,19 @@ Outstanding ones only.
 
     * Enumerator::Lazy#compact is added. [[Feature #17312]]
 
+* Module
+
+    * Module#prepend now modifies the ancestor chain if the receiver
+      already includes the argument. Module#prepend still does not
+      modify the ancestor chain if the receiver has already prepended
+      the argument. [[Bug #17423]]
+
+* Struct
+
+    * Passing only keyword arguments to Struct#initialize is warned.
+      You need to use a Hash literal to set a Hash to a first member.
+      [[Feature #16806]]
+
 ## Stdlib updates
 
 Outstanding ones only.
@@ -37,6 +50,8 @@ Excluding feature bug fixes.
 
 ### JIT
 
+* `RubyVM::MJIT` is renamed to `RubyVM::JIT`.
+
 ## Static analysis
 
 ### RBS
@@ -46,4 +61,6 @@ Excluding feature bug fixes.
 ## Miscellaneous changes
 
 
+[Feature #16806]: https://bugs.ruby-lang.org/issues/16806
 [Feature #17312]: https://bugs.ruby-lang.org/issues/17312
+[Bug #17423]: https://bugs.ruby-lang.org/issues/17423